+++ semcore/slider/__tests__/slider.vo-test.ts
@@ -14,22 +14,23 @@ test('Users can interact with Slider via VoiceOver', async ({ page, voiceOver: p
     __dirname,
     '../../../website/docs/components/slider/slider-a11y-report.md',
   );
-  const htmlContent = await e2eStandToHtml(standPath, 'en');
 
+  const htmlContent = await e2eStandToHtml(standPath, 'en');
+  await page.reload();
   await page.setContent(htmlContent);
   const { voiceOver, getReport } = await makeVoiceOverReporter(pureVoiceOver);
   await voiceOver.interact();
 
-  expect(await voiceOver.itemText()).toBe('2 slider');
+  expect(await voiceOver.itemText()).toBe('Medium slider');
   await voiceOver.interact();
   expect(await voiceOver.lastSpokenPhrase()).toBe('In slider');
   await voiceOver.press('Control+Option+ArrowLeft');
-  expect(await voiceOver.itemText()).toBe('1 slider');
+  expect(await voiceOver.itemText()).toBe('Small slider');
   await voiceOver.press('Control+Option+ArrowRight');
   await voiceOver.press('Control+Option+ArrowRight');
-  expect(await voiceOver.itemText()).toBe('3 slider');
+  expect(await voiceOver.itemText()).toBe('Big slider');
   await voiceOver.stopInteracting();
-  expect(await voiceOver.itemText()).toBe('3 slider');
+  expect(await voiceOver.itemText()).toBe('Big slider');
   expect(await voiceOver.lastSpokenPhrase()).toBe('Out of slider');
 
   const report = (await getReportHeader()) + '\n\n' + (await getReport(standPath));