+++ semcore/slider/src/index.d.ts
@@ -1,15 +1,27 @@
 import React from 'react';
 import { CProps, ReturnEl } from '@semcore/core';
-import { Box, IBoxProps } from '@semcore/flex-box';
+import { Box, IBoxProps, IFlexProps } from '@semcore/flex-box';
 
-export interface ISliderProps extends IBoxProps {
+interface ISliderContext {
+  getOptionsProps: PropGetterFn;
+  getItemProps: PropGetterFn;
+}
+interface ISliderHandlers {
+  value: (index: string | number) => void;
+}
+
+export type SliderOption<OptionValue extends stirng | number> = {
+  value: OptionValue;
+  label: React.ReactNode;
+};
+export interface ISliderProps<Value extends string | number> extends IBoxProps {
   /** Numeric value
    */
-  value?: number;
+  value?: Value;
   /** Numeric default value
    * @default 0
    */
-  defaultValue?: number;
+  defaultValue?: Value;
   /** Minimum value
    * @default 0
    */
@@ -25,16 +37,23 @@ export interface ISliderProps extends IBoxProps {
   /**
    * Handler for changing the value
    */
-  onChange?: (value: number, event: React.SyntheticEvent) => void;
+  onChange?: (value: Value, event: React.SyntheticEvent) => void;
   /**
    * Disable element
    */
   disabled?: boolean;
+
+  options?: SliderOption[];
 }
+
+interface ISliderOptionsProps extends IFlexProps {}
+interface ISliderItemProps extends IBoxProps {}
 
 declare const Slider: (<T>(props: CProps<ISliderProps & T>) => ReturnEl) & {
   Knob: typeof Box;
   Bar: typeof Box;
+  Options: <T>(props: CProps<ISliderOptionsProps & T, ISliderContext, ISliderHandlers>) => ReturnEl;
+  Item: <T>(props: CProps<ISliderItemProps & T, ISliderContext, ISliderHandlers>) => ReturnEl;
 };
 
 export default Slider;