+++ tools/jest-preset-ui/e2e-stand.ts
@@ -30,6 +30,13 @@ export const e2eStandToHtml = async (standFilePath: string, locale: string) => {
                 </I18nProvider>,
                 document.querySelector('#root')
               );
+              
+              setTimeout(() => {
+                const standReadyMarker = document.createElement('div');
+                standReadyMarker.setAttribute('id', 'AppStandReadyMarker');
+                standReadyMarker.setAttribute('hidden', true);
+                document.body.appendChild(standReadyMarker);
+              }, 500);
             `;
 
             return { contents, loader: 'tsx', resolveDir: resolveDirname(standFilePath) };
@@ -64,7 +71,7 @@ export const e2eStandToHtml = async (standFilePath: string, locale: string) => {
     .filter((file) => file.path.endsWith('.js'))
     .map((file) => file.text);
 
-  return `
+  const htmlContent = `
       <!DOCTYPE html>
       <html lang="${locale}">
         <head>
@@ -76,4 +83,6 @@ export const e2eStandToHtml = async (standFilePath: string, locale: string) => {
         </body>
       </html>
     `;
+
+  return htmlContent;
 };