+++ semcore/date-picker/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [3.3.7] - 2022-09-21
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/base-trigger` [3.0.19 ~> 3.0.20], `@semcore/icon` [2.30.3 ~> 2.30.4]).
+
+## [3.3.6] - 2022-09-21
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/base-trigger` [3.0.18 ~> 3.0.19], `@semcore/icon` [2.30.2 ~> 2.30.3]).
+
+## [3.3.5] - 2022-09-21
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/dropdown` [3.0.11 ~> 3.1.0]).
+
 ## [3.3.4] - 2022-09-20
 
 ### Changed