+++ semcore/drag-and-drop/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [2.0.18] - 2022-09-21
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/card` [4.0.18 ~> 4.0.19]).
+
+## [2.0.17] - 2022-09-21
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/card` [4.0.17 ~> 4.0.18]).
+
 ## [2.0.16] - 2022-08-31
 
 ### Changed