+++ semcore/project-create/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.1.22] - 2022-09-21
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/icon` [2.30.3 ~> 2.30.4], `@semcore/modal` [3.0.18 ~> 3.0.19]).
+
+## [5.1.21] - 2022-09-21
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/icon` [2.30.2 ~> 2.30.3], `@semcore/modal` [3.0.17 ~> 3.0.18]).
+
+## [5.1.20] - 2022-09-21
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/checkbox` [6.0.11 ~> 6.0.12]).
+
 ## [5.1.19] - 2022-09-20
 
 ### Changed